[PSym] Updates (#613)
* [PSym] symmetry: separate out explicit and symbolic symmetry tracker

Minor refactoring

* [PCover] symmetry: update explicit symmetry tracker to not use guards

* [PCover] Minor correction

* [PSym] Minor updates to logging

* [PSym] minor

* [PSym] Minor update

Do not throw stack trace for timeout/memout

* [PCover] Minor corrections

Initialize the very first step as sticky

Minor update to printing stack trace when found a bug

* [PCover] Minor update

* [PSym] refactor

* [PCompiler] Bump minimum psym version
