+++ Src/PRuntimes/PSymRuntime/src/main/java/psym/runtime/scheduler/SearchScheduler.java
@@ -1,6 +1,5 @@
 package psym.runtime.scheduler;
 
-import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -21,8 +20,6 @@ import psym.valuesummary.*;
 
 /** Represents the search scheduler */
 public abstract class SearchScheduler extends Scheduler {
-  protected transient Instant lastReportTime = Instant.now();
-
   protected SearchScheduler(PSymConfiguration config, Program p) {
     super(config, p);
   }
