+++ Src/PRuntimes/PSymRuntime/src/main/java/psym/valuesummary/SetVS.java
@@ -1,16 +1,13 @@
 package psym.valuesummary;
 
 import java.util.*;
-
 import lombok.Getter;
 import psym.runtime.machine.Machine;
 
 /** Class for set value summaries */
 public class SetVS<T extends ValueSummary<T>> implements ValueSummary<SetVS<T>> {
   @Getter
-  /**
-   * Concrete hash used for hashing in explicit-state search
-   */
+  /** Concrete hash used for hashing in explicit-state search */
   private final int concreteHash;
 
   /** The underlying set */
@@ -205,7 +202,13 @@ public class SetVS<T extends ValueSummary<T>> implements ValueSummary<SetVS<T>> 
 
   @Override
   public int computeConcreteHash() {
-    return elements.getConcreteHash();
+    int hashCode = 1;
+    List<T> allItems = new ArrayList<>(this.elements.getItems());
+    allItems.sort(Comparator.comparing(ValueSummary::getConcreteHash));
+    for (T item: allItems) {
+      hashCode = 31 * hashCode + (item == null ? 0 : item.getConcreteHash());
+    }
+    return hashCode;
   }
 
   @Override
